requirements.txt
django
web3
py-solc
eth-brownie
